refactor(packages)!: remove legacy extension paths
