+++ packages/react/src/utils/use-media-component.ts
@@ -6,8 +6,8 @@ import { useMedia } from '../player/context';
 import { useDestroy } from './use-destroy';
 
 /**
- * Create a media component (e.g. `GoogleCast`, `MuxData`) and register it with the media provided by the surrounding
- * player context.
+ * Create a media component for a custom extension and register it with the media provided by the surrounding player
+ * context.
  *
  * Instantiates the component class once, registers it when a media host is available, follows the media when it
  * changes, and destroys the component on unmount. Media that is not a media host (e.g. a plain `<video>` element)
