+++ site/src/components/docs/demos/cast-button/html/css/BasicUsage.ts
@@ -1,5 +1,5 @@
 import '@videojs/html/video/player';
-import '@videojs/html/media/google-cast';
+import '@videojs/html/extensions/google-cast';
 import '@videojs/html/ui/container';
 import '@videojs/html/media/hlsjs-video';
 import '@videojs/html/ui/cast-button';
